requirements.txt
flask
requests
python-dotenv
yt-dlp
pytube
sentence-transformers